Refactored code to multiple modules
I realized that I wanted to do more then just terminal output. So I
refactored the code into separate modules.

The API has changed to offer a Watcher main class and a TerminalWatcher
designed to do the same thing as before. In the future there will be
more Watchers (like  JsonWatcher) that can be leveraged.